requirements.txt
amqp==5.1.1
asgiref==3.6.0
attrs==23.1.0
billiard==3.6.4.0
build==0.10.0
CacheControl==0.12.11
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crashtest==0.4.1
distlib==0.3.6
Django==4.2.1
djangorestframework==3.14.0
dulwich==0.21.5
filelock==3.12.0
flower==1.2.0
html5lib==1.1
humanize==4.6.0
idna==3.4
importlib-metadata==6.6.0
installer==0.7.0
jaraco.classes==3.2.3
jsonschema==4.17.3
keyring==23.13.1
kombu==5.2.4
lockfile==0.12.2
more-itertools==9.1.0
msgpack==1.0.5
packaging==23.1
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.5.1
poetry-core==1.6.0
poetry-plugin-export==1.3.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
ptyprocess==0.7.0
pycparser==2.21
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
rapidfuzz==2.15.1
redis==4.5.5
requests==2.30.0
requests-toolbelt==1.0.0
shellingham==1.5.0.post1
six==1.16.0
sqlparse==0.4.4
tomlkit==0.11.8
tornado==6.3.2
trove-classifiers==2023.5.2
urllib3==1.26.15
vine==5.0.0
virtualenv==20.23.0
wcwidth==0.2.6
webencodings==0.5.1
xattr==0.10.1
zipp==3.15.0
